+++ src/app/components/new-cluster-summary.js
@@ -15,6 +15,8 @@ import { reads } from '@ember/object/computed';
 import { inject as service } from '@ember/service';
 
 export default Component.extend(I18n, {
+  classNames: ['new-cluster-summary'],
+
   onezoneGui: service(),
   guiUtils: service(),
   onepanelConfiguration: service(),
@@ -36,11 +38,16 @@ export default Component.extend(I18n, {
 
   isEmergencyOnepanel: reads('onepanelServer.isEmergency'),
 
+  /**
+   * @type {Ember.ComputedProperty<string>}
+   */
+  serviceType: reads('onepanelConfiguration.serviceType'),
+
   onepanelUrl: computed(function onepanelUrl() {
     if (this.get('isEmergencyOnepanel')) {
       return location.origin;
     } else {
-      return this.get('onepanelServer.emergencyOnepanelOrigin');
+      return 'https://' + this.get('onepanelServer.apiOrigin') + ':9443';
     }
   }),