+++ src/app/components/storage-item.js
@@ -4,7 +4,7 @@
  * Please put it in ``cluster-storage-table``.
  *
  * @module components/storage-item
- * @author Jakub Liput
+ * @author Jakub Liput, Michał Borzęcki
  * @copyright (C) 2017-2019 ACK CYFRONET AGH
  * @license This software is released under the MIT license cited in 'LICENSE.txt'.
  */