+++ src/app/locales/en/components/cluster-storage-add-form.js
@@ -26,4 +26,6 @@ export default {
   save: 'Save',
   add: 'Add',
   cancel: 'Cancel',
+  modifyStorageModalHeader: 'Modify storage',
+  modifyStorageModalMessage: 'Are you sure you want to modify storage details? Incorrect configuration can make your data unavailable.',
 };