+++ src/app/locales/en/components/sidebar-clusters.js
@@ -3,5 +3,6 @@ export default {
     dnsWarning: 'Some problems with your DNS configuration have been detected',
     webCertWarning: 'Some problems with your web certificate have been detected',
     cephWarning: 'Some problems with your Ceph cluster have been detected',
+    noConnectedUserWarning: 'This cluster does not have any linked Onezone user',
   },
 };