+++ src/app/locales/en/translations.js
@@ -37,8 +37,8 @@ import contentClustersProvider from './components/content-clusters-provider';
 import newCluster from './components/new-cluster';
 import modalRedirect from './components/modal-redirect';
 import clusterHostIpForm from './components/cluster-host-ip-form';
-import noAdminBox from './components/no-admin-box';
-import createAdminForm from './components/create-admin-form';
+import noEmergencyPassphraseBox from './components/no-emergency-passphrase-box';
+import firstEmergencyPassphraseForm from './components/first-emergency-passphrase-form';
 import clusterHostTable from './components/cluster-host-table';
 import clusterHostTableRow from './components/cluster-host-table-row';
 import manageClusterStorages from './components/manage-cluster-storages';
@@ -55,7 +55,7 @@ import cephClusterConfiguration from './components/ceph-cluster-configuration';
 import clusterCephStatus from './components/cluster-ceph-status';
 import clusterCephPools from './components/cluster-ceph-pools';
 import contentClustersCeph from './components/content-clusters-ceph';
-import contentClustersCredentials from './components/content-clusters-credentials';
+import contentClustersEmergencyPassphrase from './components/content-clusters-emergency-passphrase';
 import clusterNodesTile from './components/cluster-nodes-tile';
 import clusterSpacesTile from './components/cluster-spaces-tile';
 import clusterStoragesTile from './components/cluster-storages-tile';
@@ -67,6 +67,7 @@ import supportSizeInfo from './components/support-size-info';
 import emergencyWarningBar from './components/emergency-warning-bar';
 import registerOnezoneNotCompatible from './components/alerts/register-onezone-not-compatible';
 import registerOnezoneOffline from './components/alerts/register-onezone-offline';
+import contentClustersMembers from './components/content-clusters-members';
 
 import clusterIpsConfigurator from './mixins/components/cluster-ips-configurator';
 import spaceTabs from './mixins/components/space-tabs';
@@ -75,6 +76,7 @@ import guiUtils from './services/gui-utils';
 import storageActions from './services/storage-actions';
 import clusterActions from './services/cluster-actions';
 import clusterModelManager from './services/cluster-model-manager';
+import onezoneGui from './services/onezone-gui';
 
 import onedata from './routes/onedata';
 
@@ -121,8 +123,8 @@ let translations = {
     modalRedirect,
     clusterHostIpForm,
     contentClustersNodes,
-    noAdminBox,
-    createAdminForm,
+    noEmergencyPassphraseBox,
+    firstEmergencyPassphraseForm,
     clusterHostTable,
     clusterHostTableRow,
     manageClusterStorages,
@@ -141,7 +143,7 @@ let translations = {
     clusterCephStatus,
     clusterCephPools,
     contentClustersCeph,
-    contentClustersCredentials,
+    contentClustersEmergencyPassphrase,
     clusterNodesTile,
     clusterSpacesTile,
     clusterStoragesTile,
@@ -151,6 +153,7 @@ let translations = {
     spaceStorageSynchronization,
     supportSizeInfo,
     emergencyWarningBar,
+    contentClustersMembers,
     alerts: {
       registerOnezoneOffline,
       registerOnezoneNotCompatible,
@@ -167,6 +170,7 @@ let translations = {
     storageActions,
     clusterActions,
     clusterModelManager,
+    onezoneGui,
   },
   routes: {
     onedata,