+++ src/app/routes/onedata/sidebar/content.js
@@ -10,10 +10,10 @@
 import SidebarContentRoute from 'onedata-gui-common/routes/onedata/sidebar/content';
 import { inject as service } from '@ember/service';
 import { get } from '@ember/object';
-import checkImg from 'onedata-gui-common/utils/check-img';
 import { Promise } from 'rsvp';
 
 export default SidebarContentRoute.extend({
+  onepanelServer: service(),
   clusterModelManager: service(),
   navigationState: service(),
   onezoneGui: service(),
@@ -25,9 +25,11 @@ export default SidebarContentRoute.extend({
       const {
         clusterModelManager,
         onezoneGui,
+        onepanelServer,
       } = this.getProperties(
         'clusterModelManager',
-        'onezoneGui'
+        'onezoneGui',
+        'onepanelServer'
       );
 
       const currentClusterId = get(clusterModelManager, 'currentClusterProxy.id') ||
@@ -40,30 +42,25 @@ export default SidebarContentRoute.extend({
       if (clusterId !== currentClusterId) {
         if (clusterId === 'new-cluster') {
           return this.transitionTo('onedata.sidebar.content', currentClusterId);
+        } else if (onepanelServer.get('isEmergency')) {
+          return this.transitionTo('onedata.sidebar.index');
+        } else if (!get(model, 'resource')) {
+          return new Promise(() => {
+            window.location.replace(onezoneGui.getUrlInOnezone(
+              `onedata/clusters/${urlClusterId}/not-found`
+            ));
+          });
         } else {
           // If selected cluster is different than this cluster, redirect to
           // another Onepanel if possible. If Onezone is not available, then
           // redirect to main page.
           if (get(onezoneGui, 'zoneDomain')) {
-            const origin = `https://${get(model, 'resource.domain')}:9443`;
-            return checkImg(`${origin}/favicon.ico`)
-              .then(isAvailable => {
-                let redirectUrl;
-                if (isAvailable) {
-                  redirectUrl = onezoneGui.getOnepanelNavUrlInOnezone({
-                    onepanelType: get(model, 'resource.type'),
-                    clusterId,
-                    internalRoute: `/onedata/clusters/${clusterId}`,
-                  });
-                } else {
-                  redirectUrl = onezoneGui.getUrlInOnezone(
-                    `onedata/clusters/${clusterId}/endpoint-error`
-                  );
-                }
-                return new Promise(() => {
-                  window.location = redirectUrl;
-                });
-              });
+            const redirectUrl = onezoneGui.getUrlInOnezone(
+              `onedata/clusters/${clusterId}/endpoint-error`
+            );
+            return new Promise(() => {
+              window.location = redirectUrl;
+            });
           } else {
             return this.transitionTo('onedata');
           }