+++ src/app/services/gui-utils.js
@@ -34,16 +34,32 @@ export default GuiUtils.extend(
      */
     onepanelServiceType: reads('serviceType'),
 
-    guiVersion: reads('onepanelConfiguration.version'),
+    /**
+     * @override
+     */
+    softwareVersionDetails: computed(
+      'onepanelConfiguration.{version,build}',
+      function softwareVersionDetails() {
+        const onepanelConfiguration = this.get('onepanelConfiguration');
+        return {
+          serviceVersion: get(onepanelConfiguration, 'version'),
+          serviceBuildVersion: get(onepanelConfiguration, 'build'),
+        };
+      },
+    ),
 
     /**
      * Panel type: oneprovider or onezone.
      * @type {Ember.ComputedProperty<string>}
      */
-    serviceType: computed('onepanelConfiguration.serviceType', function serviceType() {
-      return this.get('onepanelServer').getClusterTypeFromUrl() ||
-        this.get('onepanelConfiguration.serviceType');
-    }),
+    serviceType: computed(
+      'onepanelConfiguration.serviceType',
+      'onepanelServer.guiContext.clusterType',
+      function serviceType() {
+        return this.get('onepanelServer.guiContext.clusterType') ||
+          this.get('onepanelConfiguration.serviceType');
+      }
+    ),
 
     /**
      * Full panel type name: Oneprovider or Onezone.
@@ -71,7 +87,7 @@ export default GuiUtils.extend(
       'onezoneGui.onezoneGuiUrl',
       function manageAccountExternalLink() {
         const onezoneGuiUrl = this.get('onezoneGui.onezoneGuiUrl');
-        return `${onezoneGuiUrl}/onedata/users`;
+        return `${onezoneGuiUrl}onedata/users`;
       }
     ),